Port TestMonthDay to scalatest
